+++ anx/provider/loadbalancer/reconciliation/reconciliation_test.go
@@ -14,8 +14,11 @@ import (
 
 	"k8s.io/apimachinery/pkg/util/rand"
 
+	"github.com/anexia-it/k8s-anexia-ccm/anx/provider/metrics"
+	"github.com/anexia-it/k8s-anexia-ccm/anx/provider/test"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/prometheus/client_golang/prometheus/testutil"
 	. "go.anx.io/go-anxcloud/pkg/api/mock/matcher"
 )
 
@@ -75,7 +78,9 @@ var _ = Describe("reconcile", func() {
 	JustBeforeEach(func() {
 		ctx := context.TODO()
 
-		r, err := New(ctx, apiClient, testClusterName, testLoadBalancerIdentifier, svcUID, externalAddresses, ports, servers)
+		metrics := metrics.NewProviderMetrics("anexia", "0.0.0-unit-tests")
+
+		r, err := New(ctx, apiClient, testClusterName, testLoadBalancerIdentifier, svcUID, externalAddresses, ports, servers, metrics)
 		Expect(err).NotTo(HaveOccurred())
 
 		recon = r.(*reconciliation)
@@ -311,10 +316,18 @@ var _ = Describe("reconcile", func() {
 			})
 
 			timeStart := time.Now()
+			timeEnd := timeStart.Add(waitTime)
 			err := recon.Reconcile()
 
 			Expect(err).NotTo(HaveOccurred())
-			Expect(time.Now()).To(BeTemporally("~", timeStart.Add(waitTime), 1*time.Second))
+			Expect(time.Now()).To(BeTemporally("~", timeEnd, 1*time.Second))
+
+			Expect(testutil.CollectAndCount(recon.metrics.ReconciliationTotalDuration, "cloud_provider_anexia_reconcile_total_duration_seconds")).To(Equal(1))
+
+			sum := test.GetHistogramSum(recon.metrics.ReconciliationTotalDuration)
+			sumDuration := time.Duration(sum * float64(time.Second))
+
+			Expect(timeStart.Add(sumDuration)).To(BeTemporally("~", timeEnd, 1*time.Second))
 		})
 
 		It("accepts the existing resources as already correct", func() {