+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Reconcile metrics
+  - Histograms
+    - cloud_provider_anexia_reconcile_total_duration_seconds
+    - cloud_provider_anexia_reconcile_create_resources_duration_seconds
+  - Counter
+    - cloud_provider_anexia_reconcile_create_errors_total
+    - cloud_provider_anexia_reconcile_delete_retries_total
+    - cloud_provider_anexia_reconcile_delete_errors_total
+    - cloud_provider_anexia_reconcile_created_total
+    - cloud_provider_anexia_reconcile_deleted_total
+    - cloud_provider_anexia_reconcile_retrieved_resources_total
+  - Gauge
+    - cloud_provider_anexia_reconcile_resources_pending
+
+### Changed
+
+* various development dependency updates
+  - Bump github.com/onsi/ginkgo/v2 from 2.1.3 to 2.8.4
+  - Bump github.com/golangci/golangci-lint from 1.45.2 to 1.51.2
+
 <!--
 Please add your changelog entry under this comment in the correct category (Security, Fixed, Added, Changed, Deprecated, Removed - in this order).
 -->